refactor: ItemsContainer.jsx를 common 폴더로 이동
